+++ types/index.d.ts
@@ -300,6 +300,8 @@ export interface VueformElement extends DefineComponent {
     [key: string]: VueformSchema;
   };
   meta: boolean;
+  forceNumbers: boolean;
+  expression: string | object;
   onAdd: Function;
   onSort: Function;
   element: object;
@@ -413,6 +415,7 @@ export interface VueformElement extends DefineComponent {
   width: string;
   top: string | number;
   bottom: string | number;
+  expressions: boolean;
   onKeydown: Function;
   onKeyup: Function;
   onKeypress: Function;
@@ -420,7 +423,6 @@ export interface VueformElement extends DefineComponent {
   onTag: Function;
   breakTags: boolean;
   showOptions: boolean;
-  forceNumbers: boolean;
   labels: object;
 
   // Computed
@@ -715,6 +717,8 @@ export interface VueformElement extends DefineComponent {
   resolveComponentName: (rowIndex: number, colIndex: number) => string;
   validateValidators: () => Promise;
   validateChildren: () => Promise;
+  shouldForceNumbers: () => boolean;
+  stringToNumber: (str: any) => number | float | string;
   add: (value: any) => number;
   handleAdd: () => void;
   refreshOrderStore: (value: Array<any>) => void;
@@ -778,8 +782,6 @@ export interface VueformElement extends DefineComponent {
   handleKeypress: (e: Event) => void;
   autosize: () => void;
   handleTag: (searchQuery: string) => void;
-  shouldForceNumbers: () => boolean;
-  stringToNumber: (str: any) => number | float | string;
 }
 
 export interface VueformSchema {
@@ -903,6 +905,8 @@ export interface VueformSchema {
     [key: string]: VueformSchema;
   };
   meta?: boolean;
+  forceNumbers?: boolean;
+  expression?: string | object;
   onAdd?: Function;
   onSort?: Function;
   element?: object;
@@ -1016,6 +1020,7 @@ export interface VueformSchema {
   width?: string;
   top?: string | number;
   bottom?: string | number;
+  expressions?: boolean;
   onKeydown?: Function;
   onKeyup?: Function;
   onKeypress?: Function;
@@ -1023,7 +1028,6 @@ export interface VueformSchema {
   onTag?: Function;
   breakTags?: boolean;
   showOptions?: boolean;
-  forceNumbers?: boolean;
   labels?: object;
   [key: string]: any;
 }
@@ -1913,6 +1917,8 @@ export interface HiddenElementProps {
   default?: string | number | object;
   id?: string;
   meta?: boolean;
+  forceNumbers?: boolean;
+  expression?: string | object;
 }
 
 export interface ListElementProps {
@@ -2786,6 +2792,7 @@ export interface StaticElementProps {
   align?: string;
   top?: string | number;
   bottom?: string | number;
+  expressions?: boolean;
 }
 
 export interface TEditorElementProps {
@@ -3126,6 +3133,7 @@ export interface TextElementProps {
   readonly?: boolean | Function | Array<any> | object;
   inputType?: string;
   forceNumbers?: boolean;
+  expression?: string | object;
   attrs?: object;
   addons?: object;
   autocomplete?: string | number;
@@ -4250,6 +4258,7 @@ export declare class Vueform implements DefineComponent {
   validate: () => Promise;
   resetValidators: () => void;
   convertFormData: (data: object) => object;
+  resolveExpression: (exp: string) => string;
   submit: () => Promise;
   send: () => Promise;
   cancel: () => void;
@@ -4262,6 +4271,7 @@ export declare class Vueform implements DefineComponent {
   el$: (path: string, elements?: object) => VueformElement | null;
   siblings$: (path: string) => void;
   initMessageBag: () => void;
+  initExpressionService: () => void;
   fire: (args?: any) => void;
   on: (event: string, callback: Function) => void;
   off: (event: string) => void;
@@ -4800,6 +4810,8 @@ export declare class CaptchaElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'beforeCreate', value: any): void;
   $emit(eventName: 'created', value: any): void;
@@ -4991,6 +5003,8 @@ export declare class CheckboxElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'beforeCreate', value: any): void;
   $emit(eventName: 'created', value: any): void;
@@ -5190,6 +5204,8 @@ export declare class CheckboxgroupElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'beforeCreate', value: any): void;
   $emit(eventName: 'created', value: any): void;
@@ -5399,6 +5415,8 @@ export declare class DateElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'beforeCreate', value: any): void;
   $emit(eventName: 'created', value: any): void;
@@ -5606,6 +5624,8 @@ export declare class DatesElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'beforeCreate', value: any): void;
   $emit(eventName: 'created', value: any): void;
@@ -5812,6 +5832,8 @@ export declare class EditorElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'blur', value: any): void;
   $emit(eventName: 'alert', value: any): void;
@@ -6041,6 +6063,8 @@ export declare class FileElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'remove', value: any): void;
   $emit(eventName: 'error', value: any): void;
@@ -6230,6 +6254,8 @@ export declare class GenericElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'beforeCreate', value: any): void;
   $emit(eventName: 'created', value: any): void;
@@ -6426,6 +6452,8 @@ export declare class GridElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'beforeCreate', value: any): void;
   $emit(eventName: 'created', value: any): void;
@@ -6606,6 +6634,8 @@ export declare class GroupElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'beforeCreate', value: any): void;
   $emit(eventName: 'created', value: any): void;
@@ -6654,6 +6684,8 @@ export declare class HiddenElement implements DefineComponent {
   default: HiddenElementProps['default'];
   id: HiddenElementProps['id'];
   meta: HiddenElementProps['meta'];
+  forceNumbers: HiddenElementProps['forceNumbers'];
+  expression: HiddenElementProps['expression'];
 
   // Computed
   isStatic: boolean;
@@ -6729,6 +6761,8 @@ export declare class HiddenElement implements DefineComponent {
   off: (event: string) => void;
   fire: (args: any) => void;
   focus: () => void;
+  shouldForceNumbers: () => boolean;
+  stringToNumber: (str: any) => number | float | string;
   validate: () => Promise;
   dirt: () => void;
   clean: () => void;
@@ -6739,6 +6773,8 @@ export declare class HiddenElement implements DefineComponent {
   reinitValidation: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'beforeCreate', value: any): void;
   $emit(eventName: 'created', value: any): void;
@@ -6950,6 +6986,8 @@ export declare class ListElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'add', value: any): void;
   $emit(eventName: 'remove', value: any): void;
@@ -7156,6 +7194,8 @@ export declare class LocationElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'beforeCreate', value: any): void;
   $emit(eventName: 'created', value: any): void;
@@ -7387,6 +7427,8 @@ export declare class MatrixElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'add', value: any): void;
   $emit(eventName: 'remove', value: any): void;
@@ -7628,6 +7670,8 @@ export declare class MultifileElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'add', value: any): void;
   $emit(eventName: 'remove', value: any): void;
@@ -7889,13 +7933,14 @@ export declare class MultiselectElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'select', value: any): void;
   $emit(eventName: 'deselect', value: any): void;
   $emit(eventName: 'search-change', value: any): void;
   $emit(eventName: 'open', value: any): void;
   $emit(eventName: 'close', value: any): void;
-  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'paste', value: any): void;
   $emit(eventName: 'beforeCreate', value: any): void;
   $emit(eventName: 'created', value: any): void;
@@ -8090,6 +8135,8 @@ export declare class ObjectElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'remove', value: any): void;
   $emit(eventName: 'beforeCreate', value: any): void;
@@ -8307,6 +8354,8 @@ export declare class PhoneElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'select', value: any): void;
   $emit(eventName: 'open', value: any): void;
@@ -8504,6 +8553,8 @@ export declare class RadioElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'beforeCreate', value: any): void;
   $emit(eventName: 'created', value: any): void;
@@ -8698,6 +8749,8 @@ export declare class RadiogroupElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'beforeCreate', value: any): void;
   $emit(eventName: 'created', value: any): void;
@@ -8950,13 +9003,14 @@ export declare class SelectElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'select', value: any): void;
   $emit(eventName: 'deselect', value: any): void;
   $emit(eventName: 'search-change', value: any): void;
   $emit(eventName: 'open', value: any): void;
   $emit(eventName: 'close', value: any): void;
-  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'paste', value: any): void;
   $emit(eventName: 'beforeCreate', value: any): void;
   $emit(eventName: 'created', value: any): void;
@@ -9275,6 +9329,8 @@ export declare class SignatureElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'beforeCreate', value: any): void;
   $emit(eventName: 'created', value: any): void;
@@ -9472,6 +9528,8 @@ export declare class SliderElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'beforeCreate', value: any): void;
   $emit(eventName: 'created', value: any): void;
@@ -9549,6 +9607,7 @@ export declare class StaticElement implements DefineComponent {
   align: StaticElementProps['align'];
   top: StaticElementProps['top'];
   bottom: StaticElementProps['bottom'];
+  expressions: StaticElementProps['expressions'];
 
   // Computed
   descriptionId: string;
@@ -9824,6 +9883,8 @@ export declare class TEditorElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'blur', value: any): void;
   $emit(eventName: 'alert', value: any): void;
@@ -10038,6 +10099,8 @@ export declare class TTextElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'blur', value: any): void;
   $emit(eventName: 'focus', value: any): void;
@@ -10255,6 +10318,8 @@ export declare class TTextareaElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'blur', value: any): void;
   $emit(eventName: 'focus', value: any): void;
@@ -10520,6 +10585,8 @@ export declare class TagsElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'select', value: any): void;
   $emit(eventName: 'deselect', value: any): void;
@@ -10527,7 +10594,6 @@ export declare class TagsElement implements DefineComponent {
   $emit(eventName: 'open', value: any): void;
   $emit(eventName: 'close', value: any): void;
   $emit(eventName: 'tag', value: any): void;
-  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'paste', value: any): void;
   $emit(eventName: 'beforeCreate', value: any): void;
   $emit(eventName: 'created', value: any): void;
@@ -10617,6 +10683,7 @@ export declare class TextElement implements DefineComponent {
   readonly: TextElementProps['readonly'];
   inputType: TextElementProps['inputType'];
   forceNumbers: TextElementProps['forceNumbers'];
+  expression: TextElementProps['expression'];
   attrs: TextElementProps['attrs'];
   addons: TextElementProps['addons'];
   autocomplete: TextElementProps['autocomplete'];
@@ -10751,6 +10818,8 @@ export declare class TextElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'blur', value: any): void;
   $emit(eventName: 'focus', value: any): void;
@@ -10966,6 +11035,8 @@ export declare class TextareaElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'blur', value: any): void;
   $emit(eventName: 'focus', value: any): void;
@@ -11167,6 +11238,8 @@ export declare class ToggleElement implements DefineComponent {
   show: () => void;
 
   //Events
+  $emit(eventName: 'reset', value: any): void;
+  $emit(eventName: 'clear', value: any): void;
   $emit(eventName: 'change', value: any): void;
   $emit(eventName: 'beforeCreate', value: any): void;
   $emit(eventName: 'created', value: any): void;